+++ src/main/java/com/aaa/api/repository/posts/dto/PostSearchForRepository.java
@@ -8,9 +8,9 @@ public class PostSearchForRepository {
 
     private final int page;
     private final int size;
-    private final long offset;
+    private final int offset;
     @Builder
-    public PostSearchForRepository(final int page, final int size, final long offset) {
+    public PostSearchForRepository(final int page, final int size, final int offset) {
         this.page = page;
         this.size = size;
         this.offset = offset;
